refactor: 댓글 컴포넌트 분리
