+++ community/viewpost.html
@@ -87,7 +87,7 @@
             </div>
         </div>
     </div>
-
+    <script type="module" src="../assets/component/comments.js"></script>
     <script type="module" src="../assets/component/Header.js"></script>
     <script type="module" src="viewpost.js"></script>
 </body>
